moving constants  to a new file
